scripts/requirements_scripts.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
psycopg2-binary>=2.9.9
